+++ data/tasks.txt
@@ -1,3 +1,2 @@
-E | 0 | u | from 2024-01-02 to 2024-02-01
-E | 0 | ss | from 2023-01-01 to 2024-02-02
-D | 0 | dsdhi | by2024-02-03
+T | 0 | 999999
+D | 0 | 22222 | by2024-03-02
